ver. 1.0.3/20180918

+++ hois_back/src/main/java/ee/hitsa/ois/web/dto/timetable/LessonPlanSearchDto.java
@@ -1,13 +1,15 @@
 package ee.hitsa.ois.web.dto.timetable;
 
+import java.math.BigDecimal;
+
 public class LessonPlanSearchDto {
 
     private final Long id;
     private final String studentGroup;
     private final String curriculumVersion;
-    private final Long plannedHours;
+    private final BigDecimal plannedHours;
 
-    public LessonPlanSearchDto(Long id, String studentGroup, String curriculumVersion, Long plannedHours) {
+    public LessonPlanSearchDto(Long id, String studentGroup, String curriculumVersion, BigDecimal plannedHours) {
         this.id = id;
         this.studentGroup = studentGroup;
         this.curriculumVersion = curriculumVersion;
@@ -26,7 +28,7 @@ public class LessonPlanSearchDto {
         return curriculumVersion;
     }
 
-    public Long getPlannedHours() {
+    public BigDecimal getPlannedHours() {
         return plannedHours;
     }
